Use only one posthook for installing Istio
