+++ client/SHA256SUMS
@@ -1,1 +1,1 @@
-fdb101513b3c992b24ca7d96b0075f025d899f5d3964a4409cc9533bbcf13a30  docs/openapi/pipeline.yaml
+0dddc4826a74c82c0ecc2368d8199cf536712f8ee12b7d77d801c9e11efe6913  docs/openapi/pipeline.yaml
