+++ client/docs/PostHooks.md
@@ -5,7 +5,6 @@ Name | Type | Description | Notes
 ------------ | ------------- | ------------- | -------------
 **InstallLogging** | [**LoggingPostHookInstallLogging**](LoggingPostHook_InstallLogging.md) |  | [optional] 
 **PostHookFunctionName** | [**map[string]interface{}**](.md) |  | [optional] 
-**InstallIstioOperator** | [**map[string]interface{}**](.md) |  | [optional] 
 **InstallServiceMesh** | [**ServiceMeshPostHookInstallServiceMesh**](ServiceMeshPostHook_InstallServiceMesh.md) |  | [optional] 
 
 [[Back to Model list]](../README.md#documentation-for-models) [[Back to API list]](../README.md#documentation-for-api-endpoints) [[Back to README]](../README.md)
