+++ client/docs/ServiceMeshPostHook.md
@@ -3,7 +3,6 @@
 ## Properties
 Name | Type | Description | Notes
 ------------ | ------------- | ------------- | -------------
-**InstallIstioOperator** | [**map[string]interface{}**](.md) |  | [optional] 
 **InstallServiceMesh** | [**ServiceMeshPostHookInstallServiceMesh**](ServiceMeshPostHook_InstallServiceMesh.md) |  | [optional] 
 
 [[Back to Model list]](../README.md#documentation-for-models) [[Back to API list]](../README.md#documentation-for-api-endpoints) [[Back to README]](../README.md)
